:sparkles: Supporting vanue and address
New event format: Support of venue and address instead of title and location. The address will be used to automatically add a “Show on map” link to the output. The old format is still supported.

In the past, the title field was used to specify the company or venue and the location field was used to specify the city and country. The new field now clearly describe their contents.

For today we do not migrate the old events. It may produce incorrect addresses for the map links and a manual cleanup is very much work. For now it is absolutely okay to support the old format, too.