Parse unique locations from events
We use all upcoming and all events within the last 52 weeks to figure
out 'active' Hackergarten locations.

The venue of the event is used as sponsor. Later on we can add all
different venues of a location to the sponsors list.

We parse the link of the meetup group from the event link with no
fallback.

TODOs:
- The citiy in the title should be replaced by the city of the address.
- Long and Lat should be determined using the address and the
openstreetmap api (used in hero-geo-location npm package).
- Use modern JS to simplify the code.